upgrade: Next.js 15, TypeScript, ESLint flat config, async params
