--- frontend/.eslintrc.json
@@ -1,9 +1,0 @@
-{
-  "extends": [
-    "next/core-web-vitals",
-    "plugin:react-hooks/recommended"
-  ],
-  "rules": {
-    "@next/next/no-img-element": "off"
-  }
-}
