+++ frontend/eslint.config.mjs
@@ -1,0 +1,16 @@
+import { FlatCompat } from "@eslint/eslintrc"
+
+const compat = new FlatCompat({
+  baseDirectory: import.meta.dirname,
+})
+
+const eslintConfig = [
+  ...compat.extends("next/core-web-vitals"),
+  {
+    rules: {
+      "@next/next/no-img-element": "off",
+    },
+  },
+]
+
+export default eslintConfig
