+++ frontend/next.config.mjs
@@ -11,4 +11,4 @@ const nextConfig = {
     }
 }
 
-module.exports = nextConfig
+export default nextConfig
