+++ frontend/package.json
@@ -1,14 +1,8 @@
 {
   "name": "frontend",
   "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
   "dependencies": {
+    "@eslint/eslintrc": "^3.3.5",
     "@headlessui/react": "^1.7.19",
     "@heroicons/react": "^2.2.0",
     "@hookform/resolvers": "^3.10.0",
@@ -16,36 +10,40 @@
     "@tanstack/react-query": "^4.44.0",
     "@tanstack/react-query-devtools": "^4.44.0",
     "@types/lodash": "^4.17.24",
-    "@types/node": "20.6.3",
-    "@types/react": "18.2.22",
-    "@types/react-dom": "18.2.7",
+    "@types/node": "22.15.3",
+    "@types/react": "18.3.18",
+    "@types/react-dom": "18.3.5",
     "@types/react-speech-recognition": "^3.9.6",
     "autoprefixer": "10.4.16",
     "awesome-debounce-promise": "^2.1.0",
     "axios": "^1.18.0",
-    "eslint": "8.49.0",
-    "eslint-config-next": "13.5.2",
+    "eslint": "^9",
+    "eslint-config-next": "15.5.19",
     "eslint-plugin-react": "^7.37.5",
     "framer-motion": "^10.18.0",
     "lodash": "^4.18.1",
-    "next": "^13.5.11",
+    "next": "15.5.19",
     "next-auth": "^4.24.14",
     "nextjs-toploader": "^1.6.12",
     "openai": "^4.104.0",
     "postcss": "8.4.30",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
     "react-hook-form": "^7.80.0",
     "react-select": "^5.10.2",
     "react-speech-recognition": "^3.10.0",
     "react-toastify": "^9.1.3",
     "tailwindcss": "3.3.3",
-    "typescript": "5.2.2",
+    "typescript": "5.8.3",
     "use-constant": "^1.1.1",
     "zod": "^3.25.76",
     "zustand": "^4.5.7"
   },
-  "devDependencies": {
-    "eslint-plugin-react-hooks": "^4.6.2"
+  "private": true,
+  "scripts": {
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "lint": "eslint . --ext .ts,.tsx,.js,.jsx"
   }
 }
