+++ frontend/tailwind.config.ts
@@ -15,9 +15,5 @@ const config: Config = {
       },
     },
   },
-  plugins: [
-    require('@tailwindcss/line-clamp'),
-
-  ],
 }
 export default config
